refactor(world): publish canonical value contracts
