+++ docs/reference/python/configuration.md
@@ -43,7 +43,6 @@
 
 | Field | Type | Default | Description |
 | --- | --- | --- | --- |
-| `run_id` | `str \| UUID` | `generated by uuid7` | Stable identifier for this run, generated when omitted. |
 | `num_steps` | `int` | `1` | Number of ticks to execute. |
 | `debug` | `bool` | `False` | Emit per-tick diagnostic panels. |
 | `show_rows` | `int` | `8` | Maximum rows shown in each debug snapshot; zero disables snapshots. |
@@ -57,7 +56,6 @@
 | --- | --- | --- | --- |
 | `world_id` | `str \| UUID \| None` | `generated by uuid7` | Unique identifier for the world. Auto-generated if omitted. |
 | `name` | `str \| None` | `None` | Human-readable alias for the world |
-| `run_id` | `str \| None` | `None` | Active run identifier |
 | `tick` | `int` | `0` | Current simulation tick |
 | `next_entity_id` | `int` | `1` | Next entity ID to assign |
 | `entity2sig` | `dict[int, tuple]` | `generated by dict` | Entity ID → archetype signature mapping |
