+++ docs/reference/python/runtime-models.md
@@ -4,107 +4,107 @@
 
 **Recommended API.** Configuration, result, and introspection records returned by world operations.
 
-::: archetype.app.models.ProcessorInfo
+::: archetype.world.models.ProcessorInfo
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `qualname` | `str` | `required` | Qualified processor class name. |
-| `priority` | `int` | `0` | Execution priority; lower values run first. |
-| `components` | `tuple[str, ...]` | `generated by tuple` | Qualified component names required by the processor. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `qualname` | `str` | `required` |
+| `priority` | `int` | `0` |
+| `components` | `tuple[str, ...]` | `()` |
 
-::: archetype.app.models.HookInfo
+::: archetype.world.models.HookInfo
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `event_type` | `str` | `required` | Qualified lifecycle-event class name. |
-| `handler_qualname` | `str` | `required` | Qualified handler name. |
-| `mode` | `str` | `'blocking'` | Execution mode: blocking or spawn. |
-| `handle_id` | `int` | `required` | Handle identifier used to remove the hook. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `event_type` | `str` | `required` |
+| `handler_qualname` | `str` | `required` |
+| `mode` | `str` | `'blocking'` |
+| `handle_id` | `int` | `required` |
 
-::: archetype.app.models.ResourceInfo
+::: archetype.world.models.ResourceInfo
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `qualname` | `str` | `required` | Qualified resource class name. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `qualname` | `str` | `required` |
 
-::: archetype.app.models.WorldInfo
+::: archetype.world.models.WorldInfo
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `world_id` | `str \| UUID` | `required` | Durable world identifier. |
-| `name` | `str \| None` | `None` | Human-readable world name. |
-| `tick` | `int` | `0` | Next tick to execute. |
-| `run_id` | `str \| UUID \| None` | `None` | Identifier of the active or most recent run. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `world_id` | `str \| UUID` | `required` |
+| `name` | `str \| None` | `None` |
+| `tick` | `int` | `0` |
+| `run_id` | `str \| UUID` | `required` |
 
-::: archetype.app.models.RunResult
+::: archetype.world.models.RunResult
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `run_id` | `str \| UUID` | `required` | Identifier of the completed run. |
-| `world_id` | `str \| UUID` | `required` | World advanced by the run. |
-| `ticks_completed` | `int` | `0` | Number of ticks executed. |
-| `commands_applied` | `int` | `0` | Number of queued commands applied during the run. |
-| `final_tick` | `int` | `0` | World tick after the run completed. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `run_id` | `str \| UUID` | `required` |
+| `world_id` | `str \| UUID` | `required` |
+| `ticks_completed` | `int` | `0` |
+| `commands_applied` | `int` | `0` |
+| `final_tick` | `int` | `0` |
 
-::: archetype.app.models.EpisodeConfig
+::: archetype.world.models.EpisodeConfig
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `episode_id` | `str \| UUID` | `generated by uuid7` | Stable identifier for this episode. |
-| `run_config` | `RunConfig` | `generated by RunConfig` | Tick execution options. |
-| `max_steps` | `int` | `1000` | Maximum ticks before stopping. |
-| `terminal_component` | `Any \| None` | `None` | Component type used for structural or value termination. |
-| `terminal_field` | `str \| None` | `None` | Boolean field tested on the terminal component. |
-| `terminal_all` | `bool` | `True` | Require every matching entity when testing a field. |
-| `termination` | `Any \| None` | `None` | Optional callable termination predicate. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `episode_id` | `str \| UUID` | `generated by uuid7` |
+| `run_config` | `RunConfig` | `generated by RunConfig` |
+| `max_steps` | `int` | `1000` |
+| `terminal_component` | `type[Component] \| None` | `None` |
+| `terminal_field` | `str \| None` | `None` |
+| `terminal_all` | `bool` | `True` |
+| `termination` | `Callable[[Any], bool] \| None` | `None` |
 
-::: archetype.app.models.EpisodeResult
+::: archetype.world.models.EpisodeResult
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `episode_id` | `str \| UUID` | `required` | Episode identifier. |
-| `world_id` | `str \| UUID` | `required` | Forked world used by the episode. |
-| `run_id` | `str \| UUID \| None` | `None` | Episode run identifier. |
-| `start_tick` | `int` | `0` | World tick at episode start. |
-| `final_tick` | `int` | `0` | World tick at episode completion. |
-| `terminated` | `bool` | `False` | Whether a termination condition stopped the episode. |
-| `duration_steps` | `int` | `0` | Number of ticks executed. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `episode_id` | `str \| UUID` | `required` |
+| `world_id` | `str \| UUID` | `required` |
+| `run_id` | `str \| UUID` | `required` |
+| `start_tick` | `int` | `0` |
+| `final_tick` | `int` | `0` |
+| `terminated` | `bool` | `False` |
+| `duration_steps` | `int` | `0` |
 
-::: archetype.app.models.RolloutConfig
+::: archetype.world.models.RolloutConfig
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `rollout_id` | `str \| UUID` | `generated by uuid7` | Stable identifier for this rollout. |
-| `episode_config` | `EpisodeConfig` | `generated by EpisodeConfig` | Configuration shared by each episode. |
-| `num_episodes` | `int` | `1` | Number of episode forks to run. |
-| `parallel` | `bool` | `False` | Run episode forks concurrently. |
-| `name_prefix` | `str` | `'ep'` | Name prefix for episode worlds. |
-| `destroy_forks_on_complete` | `bool` | `False` | Destroy live episode worlds after collecting results. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `rollout_id` | `str \| UUID` | `generated by uuid7` |
+| `episode_config` | `EpisodeConfig` | `generated by EpisodeConfig` |
+| `num_episodes` | `int` | `1` |
+| `parallel` | `bool` | `False` |
+| `name_prefix` | `str` | `'ep'` |
+| `destroy_forks_on_complete` | `bool` | `False` |
 
-::: archetype.app.models.RolloutResult
+::: archetype.world.models.RolloutResult
     options:
       members: false
 
-| Field | Type | Default | Description |
-| --- | --- | --- | --- |
-| `rollout_id` | `str \| UUID` | `required` | Rollout identifier. |
-| `base_world_id` | `str \| UUID` | `required` | World forked for each episode. |
-| `episodes` | `list[EpisodeResult]` | `generated by list` | Results in episode order. |
-| `num_episodes` | `int` | `0` | Number of completed episodes. |
-| `total_duration_steps` | `int` | `0` | Total ticks executed across all episodes. |
+| Field | Type | Default |
+| --- | --- | --- |
+| `rollout_id` | `str \| UUID` | `required` |
+| `base_world_id` | `str \| UUID` | `required` |
+| `episodes` | `tuple[EpisodeResult, ...]` | `()` |
+| `num_episodes` | `int` | `0` |
+| `total_duration_steps` | `int` | `0` |
