+++ scripts/generate_python_api_docs.py
@@ -233,7 +233,7 @@ SUPPLEMENTAL: dict[str, tuple[str, str]] = {
     "ActorCtx": ("archetype.app.gateway.auth", "ActorCtx"),
     "HookEvent": ("archetype.core.hooks", "HookEvent"),
     "HookHandle": ("archetype.core.hooks", "HookHandle"),
-    "HookInfo": ("archetype.app.models", "HookInfo"),
+    "HookInfo": ("archetype.world.models", "HookInfo"),
     "ClaudeTranscriptSource": (
         "archetype.missions.trajectories",
         "ClaudeTranscriptSource",
@@ -259,8 +259,8 @@ SUPPLEMENTAL: dict[str, tuple[str, str]] = {
         "archetype.physical_ai.optimization",
         "optimize_instruction",
     ),
-    "ProcessorInfo": ("archetype.app.models", "ProcessorInfo"),
-    "ResourceInfo": ("archetype.app.models", "ResourceInfo"),
+    "ProcessorInfo": ("archetype.world.models", "ProcessorInfo"),
+    "ResourceInfo": ("archetype.world.models", "ResourceInfo"),
     "StorageBackend": ("archetype.core.config", "StorageBackend"),
     "PreTick": ("archetype.core.hooks", "PreTick"),
     "PostTick": ("archetype.core.hooks", "PostTick"),
