+++ src/archetype/runtime/runtime.py
@@ -26,7 +26,6 @@ from uuid_utils import UUID
 from archetype._logging import configure_host_observability
 from archetype.app.application.interfaces import iRuntimeApplication
 from archetype.app.container import ServiceContainer
-from archetype.app.models import WorldInfo
 from archetype.artifacts.contracts import ArtifactStoreConfig
 from archetype.core.config import CacheConfig, StorageConfig
 from archetype.core.hooks import HookEvent
@@ -36,6 +35,7 @@ from archetype.runtime.world import (
     SyncRuntimeWorld,
     _RuntimeWorldState,
 )
+from archetype.world.models import WorldInfo
 
 if TYPE_CHECKING:
     from archetype.missions.contracts import AgentMissionConfig
